# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Circl social networking API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User profiles and discovery
  - name: connections
    description: Connection requests and matches
  - name: conversations
    description: Chat conversations and messages
  - name: notifications
    description: User notifications
  - name: dashboard
    description: Dashboard stats
  - name: blocks
    description: User blocking
  - name: reports
    description: Content and user reports

security:
  - bearerAuth: []

paths:
  /healthz:
    get:
      security: []
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users/me:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    delete:
      operationId: deleteMyAccount
      tags: [users]
      summary: Soft-delete the current user account (scheduled for permanent deletion in 30 days)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAccountResult"

  /users:
    get:
      operationId: discoverUsers
      tags: [users]
      summary: Discover users with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: university
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscoverResult"

  /users/{id}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get a user's profile by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found

  /connections:
    get:
      operationId: listConnections
      tags: [connections]
      summary: List connections for current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, rejected, passed]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Connection"
    post:
      operationId: createConnection
      tags: [connections]
      summary: Send a connect or pass request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"

  /connections/{id}:
    patch:
      operationId: updateConnection
      tags: [connections]
      summary: Accept or reject a connection request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionPatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"

  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      summary: List conversations for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"

  /conversations/{id}/messages:
    get:
      operationId: getMessages
      tags: [conversations]
      summary: Get messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [conversations]
      summary: Send a message in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"

  /notifications/read-all:
    patch:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadAllResult"

  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  /users/me/blocks:
    get:
      operationId: listBlockedUsers
      tags: [blocks]
      summary: List users blocked by the current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockedUser"

  /users/{id}/block:
    post:
      operationId: blockUser
      tags: [blocks]
      summary: Block a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockResult"
    delete:
      operationId: unblockUser
      tags: [blocks]
      summary: Unblock a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockResult"

  /reports:
    post:
      operationId: createReport
      tags: [reports]
      summary: Report a user, activity, or message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics and recent activity
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    UserProfile:
      type: object
      required: [id, name, university, interests, lookingFor, availability, isOnline, profileCompletion, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        university:
          type: string
        bio:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
        section:
          type: ["string", "null"]
        studentType:
          type: ["string", "null"]
          enum: [hostel, day_scholar, null]
        githubUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        interests:
          type: array
          items:
            type: string
        lookingFor:
          type: array
          items:
            type: string
        availability:
          type: array
          items:
            type: string
        goals:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isOnline:
          type: boolean
        profileCompletion:
          type: integer
          minimum: 0
          maximum: 100
        createdAt:
          type: string

    DiscoverResult:
      type: object
      required: [user, compatibilityScore]
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        compatibilityScore:
          type: integer

    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        department:
          type: string
        year:
          type: string
        section:
          type: string
        studentType:
          type: string
          enum: [hostel, day_scholar]
        githubUrl:
          type: string
        linkedinUrl:
          type: string
        interests:
          type: array
          items:
            type: string
        lookingFor:
          type: array
          items:
            type: string
        availability:
          type: array
          items:
            type: string
        goals:
          type: string
        avatarUrl:
          type: string

    Connection:
      type: object
      required: [id, fromUserId, toUserId, status, createdAt]
      properties:
        id:
          type: integer
        fromUserId:
          type: integer
        toUserId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, rejected, passed]
        fromUser:
          $ref: "#/components/schemas/UserProfile"
        toUser:
          $ref: "#/components/schemas/UserProfile"
        createdAt:
          type: string

    ConnectionInput:
      type: object
      required: [toUserId, action]
      properties:
        toUserId:
          type: integer
        action:
          type: string
          enum: [connect, pass]

    ConnectionPatch:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, rejected]

    Conversation:
      type: object
      required: [id, otherUser, unreadCount, updatedAt]
      properties:
        id:
          type: integer
        otherUser:
          $ref: "#/components/schemas/UserProfile"
        lastMessage:
          type: ["string", "null"]
        unreadCount:
          type: integer
        updatedAt:
          type: string

    Message:
      type: object
      required: [id, conversationId, senderId, content, isRead, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        content:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string

    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    Notification:
      type: object
      required: [id, type, message, isRead, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [connection_request, connection_accepted, message, match]
        message:
          type: string
        isRead:
          type: boolean
        actorId:
          type: ["integer", "null"]
        actor:
          $ref: "#/components/schemas/UserProfile"
        createdAt:
          type: string

    ReadAllResult:
      type: object
      required: [count]
      properties:
        count:
          type: integer

    DashboardStats:
      type: object
      required: [totalMatches, pendingRequests, newMessages, profileViews, recentMatches, activityByCategory]
      properties:
        totalMatches:
          type: integer
        pendingRequests:
          type: integer
        newMessages:
          type: integer
        profileViews:
          type: integer
        recentMatches:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        activityByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryActivity"

    CategoryActivity:
      type: object
      required: [category, count]
      properties:
        category:
          type: string
        count:
          type: integer

    DeleteAccountResult:
      type: object
      required: [message]
      properties:
        message:
          type: string

    BlockResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    BlockedUser:
      type: object
      required: [id, blockedUser, createdAt]
      properties:
        id:
          type: integer
        blockedUser:
          $ref: "#/components/schemas/UserProfile"
        createdAt:
          type: string

    ReportInput:
      type: object
      required: [targetType, targetId, reason]
      properties:
        targetType:
          type: string
          enum: [user, activity, message]
        targetId:
          type: integer
        reason:
          type: string
        description:
          type: string

    Report:
      type: object
      required: [id, reporterId, targetType, targetId, reason, status, createdAt]
      properties:
        id:
          type: integer
        reporterId:
          type: integer
        targetType:
          type: string
        targetId:
          type: integer
        reason:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string