# lib/api-zod/src/generated/types/blockedUser.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Circl social networking API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from './userProfile';

export interface BlockedUser {
  id: number;
  blockedUser: UserProfile;
  createdAt: string;
}