# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Circl social networking API
 * OpenAPI spec version: 0.1.0
 */

export interface Report {
  id: number;
  reporterId: number;
  targetType: string;
  targetId: number;
  reason: string;
  /** @nullable */
  description?: string | null;
  status: string;
  createdAt: string;
}