# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Circl social networking API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportInputTargetType } from './reportInputTargetType';

export interface ReportInput {
  targetType: ReportInputTargetType;
  targetId: number;
  reason: string;
  description?: string;
}